refactor: split OAuth discovery into own file, fix test race, add docs
- Extract OAuth cloud ID discovery logic from jira.go into
  oauth_discovery.go (jira.go: 402→231 lines)
- Move OAuth tests to oauth_discovery_test.go with race-safe
  sync/atomic counter (jira_test.go: 652→393 lines)
- Add Atlassian OAuth setup section to docs/oauth-setup.md
- Update docs/creating-connectors.md to reference Jira as dual
  OAuth/basic auth example
